feat(checkout): CHECKOUT-6621 Refactor custom fixtures
