+++ tests/_fixture/index.ts
@@ -1,0 +1,1 @@
+export { expect, test } from './CustomFixtures';
