+++ tests/sampleTests/template.ts
@@ -1,16 +1,16 @@
-import { test } from '../testService';
+import { test } from '../';
 
 test.describe('xxxxxxxxxxxxxxxxxxxxxxx', () => {
 
-    test.beforeEach(async ({ testService }) => {
+    test.beforeEach(async ({ paymentStep }) => {
         // Setup HAR and checkout environment.
-        await testService.gotoPaymentStep({ storeURL: 'https://my-dev-store-745516528.store.bcdev', HAR: 'sampleName' });
+        await paymentStep.goto({ storeURL: 'https://my-dev-store-745516528.store.bcdev', HAR: 'sampleName' });
     });
 
-    test('xxxxxxxxxxxxxxxxxxxxxxx I', async ({testService, page}) => {
+    test('xxxxxxxxxxxxxxxxxxxxxxx I', async ({paymentStep, page}) => {
         // Optional: Serve additional static files with Playwright. Mock API endpoints via Playwright.
         // Only execute in the replay mode
-        if (testService.isReplay) {
+        if (paymentStep.isReplay) {
             // await page.route('**/checkout/payment/hosted-field?**', route => route.fulfill( {status: 200, path: './tests/sampleTests/_support/hostedField.html' } ));
             // await page.route('**/api/public/v1/orders/payments', route => route.fulfill( {status: 200, contentType: 'application/json', body: JSON.stringify(submitPaymentResult)} ));
         }
@@ -23,22 +23,22 @@ test.describe('xxxxxxxxxxxxxxxxxxxxxxx', () => {
         // await page.frameLocator('#bigpaypay-ccNumber iframe').locator('[aria-label="Credit Card Number"]').click();
 
         // Assertions
-        await testService.shouldSeeOrderConfirmation();
+        await paymentStep.shouldSeeOrderConfirmation();
     });
 
-    test('xxxxxxxxxxxxxxxxxxxxxxx II', async ({testService, page}) => {
+    test('xxxxxxxxxxxxxxxxxxxxxxx II', async ({paymentStep, page}) => {
         // Playwright action scripts
         await page.pause();
 
         // Assertions
-        await testService.shouldSeeOrderConfirmation();
+        await paymentStep.shouldSeeOrderConfirmation();
     });
 
-    test('xxxxxxxxxxxxxxxxxxxxxxx III', async ({testService, page}) => {
+    test('xxxxxxxxxxxxxxxxxxxxxxx III', async ({paymentStep, page}) => {
         // Playwright action scripts
         await page.pause();
 
         // Assertions
-        await testService.shouldSeeOrderConfirmation();
+        await paymentStep.shouldSeeOrderConfirmation();
     });
 });
